CategroyController created, set of ProductController, templates HTML also created

+++ src/main/java/com/sadev/app/repository/CategoryRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.sadev.app.model.Category;
 
 @Repository
-public interface CategoryRepository extends CrudRepository<Category, Long> {
+public interface CategoryRepository extends CrudRepository<Category, Integer> {
 
 }
